requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
datasets==4.5.0
dill==0.3.8
distro==1.9.0
docker==7.1.0
dynaconf==3.2.12
entrypoints==0.4
evidently==0.4.30
fake-review-detector==1.0.0
Faker==40.4.0
fastapi==0.110.0
filelock==3.24.3
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.46
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
huggingface_hub==1.4.1
humanfriendly==10.0
idna==3.11
importlib_metadata==7.2.1
iniconfig==2.3.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
litestar==2.21.0
litestar-htmx==0.5.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mlflow==2.11.0
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.1
multipart==1.3.0
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.16.0
networkx==3.6.1
nltk==3.9.2
numpy==1.26.0
onnxruntime==1.17.0
packaging==23.2
pandas==2.2.0
patsy==1.0.2
pillow==12.1.1
plotly==6.5.2
pluggy==1.6.0
polyfactory==3.2.0
propcache==0.4.1
protobuf==4.25.8
psutil==7.2.2
pyarrow==23.0.1
pyarrow-hotfix==0.7
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.19.2
pyparsing==3.3.2
pyreadline3==3.5.4
pytest==8.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
pywin32==311
PyYAML==6.0.3
querystring-parser==1.2.4
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rich-click==1.9.7
safetensors==0.7.0
scikit-learn==1.4.0
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlparse==0.5.5
starlette==0.36.3
statsmodels==0.14.6
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.27.0
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
